Modules fetchAndRenderComments.js and addComments.js are added
